feat: Editable payee + refactor
Apart from continuing work on the transaction edit modal i've also refactored and broken out logic for the shortcuts and keybindings into it's own context provider and hooks.

This way all keybindings dont need to live in the page component any more, the edit modal can define it's own keybindings – seperation of concern.

It was a bit tricky but the way it works is that the hooks consumed does the keypress logic and the only data propagated back to the provider is what is need to visually present the shortcuts.

Also includes removing experimentational logic for selecting a visible transaction on arrow key press.